Fixing Last Error before Adding Schedules

+++ kernels/temp_Range_kernel.cl
@@ -16,7 +16,7 @@ int inRange(float value, float min, float max){
  *
  *      wLoad:      Wert 1 für voll parallele Ausführung, Wert MAX_ITEM_SIZE_PER_WORKGROUP für vollständig sequenzielle Ausführung.
  */
-__kernel void tempRange(global const unsigned float* input, const float min, const float max, global int* result, const int wLoad) {
+__kernel void tempRange(global const float* input, const float min, const float max, global int* result, const int wLoad) {
     int gid = get_global_id(0) * wLoad;
 
     for (int i = 0; i < wLoad; ++i){
@@ -27,7 +27,7 @@ __kernel void tempRange(global const unsigned float* input, const float min, con
 } 
 
 //Mit Array Min,Max
-__kernel void temp1(global const unsigned float* input, global const float* min, global const float*max, global int* result, const int wLoad) {
+__kernel void temp1(global const float* input, global const float* min, global const float*max, global int* result, const int wLoad) {
     int gid = get_global_id(0) * wLoad;
 
     for (int i = 0; i < wLoad; ++i){
